requirements.txt
click==8.1.8
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
pyprojroot==0.3.0
PyYAML==6.0.2
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
typer==0.15.2
typing_extensions==4.13.2
